+++ src/pages/mypage/passwordCheck.tsx
@@ -1,18 +1,43 @@
 'use client';
 import { useEffect, useState } from 'react';
 import styles from './styles/mypage.module.scss';
-import { usePasswordChange } from './passwordChange';
 import Button from '@/components/common/button/myPageButton/MypageButton';
 import Input from '@/components/common/Input';
 import buttonStyles from '@/components/common/button/myPageButton/myPageButton.module.scss';
-import CreateColumnModal from '@/components/modal/CreateColumnModal';
+import { useChangePasswordMutation } from '@/api/auth/auth.query';
+import { useRouter } from 'next/navigation';
+import OneButtonModal from '@/components/modal/OneButtonModal';
 
 export default function PasswordCheck() {
   const [currentPassword, setCurrentPassword] = useState('');
   const [newPassword, setNewPassword] = useState('');
   const [confirmPassword, setConfirmPassword] = useState('');
   const [isFormValid, setIsFormValid] = useState(false);
-  const { changePassword } = usePasswordChange();
+  const [isMismatch, setIsMismatch] = useState(false);
+  const [isTouched, setIsTouched] = useState(false);
+  const router = useRouter();
+  const { mutate } = useChangePasswordMutation();
+  const [isModalOpen, setIsModalOpen] = useState(false);
+
+  const changePassword = (currentPassword: string, newPassword: string) => {
+    console.log('비밀번호 변경 요청 시작');
+
+    mutate(
+      {
+        password: currentPassword,
+        newPassword: newPassword,
+      },
+      {
+        onSuccess: () => {
+          alert('변경 성공!');
+          router.refresh();
+        },
+        onError: (error: any) => {
+          setIsModalOpen(true);
+        },
+      },
+    );
+  };
 
   useEffect(() => {
     if (currentPassword && newPassword && confirmPassword) {
@@ -26,9 +51,9 @@ export default function PasswordCheck() {
     if (!isFormValid) return;
 
     if (newPassword !== confirmPassword) {
-      alert('새 비밀번호가 일치하지 않습니다.');
       return;
     }
+
     changePassword(currentPassword, newPassword);
   };
 
@@ -70,8 +95,15 @@ export default function PasswordCheck() {
                   placeholder="새 비밀번호 확인"
                   value={confirmPassword}
                   onChange={(e) => setConfirmPassword(e.target.value)}
-                  className={styles.inputBox}
+                  onBlur={() => {
+                    setIsTouched(true);
+                    setIsMismatch(newPassword !== confirmPassword);
+                  }}
+                  className={`${styles.inputBox} ${isTouched && isMismatch ? styles.errorInput : ''}`}
                 />
+                {isTouched && isMismatch && (
+                  <p className={styles.errorText}>새 비밀번호가 일치하지 않습니다</p>
+                )}
               </div>
             </div>
           </form>
@@ -83,7 +115,9 @@ export default function PasswordCheck() {
             변경
           </Button>
         </div>
-        <CreateColumnModal />
+        {isModalOpen && (
+          <OneButtonModal message="실패했습니다." onClose={() => setIsModalOpen(false)} />
+        )}
       </div>
     </div>
   );
