+++ src/pages/mypage/styles/mypage.module.scss
@@ -98,3 +98,18 @@
 .passwordInputBox {
   margin-top: -30px;
 }
+.inputBox {
+  border: 1px solid #ccc;
+  padding: 8px;
+  border-radius: 4px;
+}
+
+.errorInput {
+  border-color: red;
+}
+
+.errorText {
+  color: red;
+  font-size: 0.875rem;
+  margin-top: 4px;
+}
